refactor(runners): extract error classification into api/runners/errors (Phase 1)
First step of the mission_runner.rs decomposition (18.6k lines): the
provider/CLI error predicates and their ASCII matching helpers move
verbatim to src/api/runners/errors.rs. mission_runner.rs re-exports them
so all call sites and existing tests keep their paths.

Adds golden marker tests pinning classification of real incident strings
(auth 401/refresh-token-reuse, 429/529/quota phrasings, capacity
rejections, success-path error payloads) so future edits to the marker
tables are caught as intentional behavior changes.

No behavior change.
