+++ src/api/mod.rs
@@ -49,6 +49,7 @@ mod providers;
 pub(crate) mod proxy;
 mod proxy_keys;
 mod routes;
+pub(crate) mod runners;
 pub mod secrets;
 pub mod settings;
 pub mod system;
